chore(ts): move unit test fs mock to ts
